Bump MSRV to 1.60 for csv dependency (#665)

+++ src/plot/plotters_backend/pdf.rs
@@ -38,7 +38,7 @@ pub(crate) fn pdf_comparison_figure(
     let y_range = data::fitting_range(base_ys.iter().chain(ys.iter()));
 
     let size = size.unwrap_or(SIZE);
-    let root_area = SVGBackend::new(&path, (size.0 as u32, size.1 as u32)).into_drawing_area();
+    let root_area = SVGBackend::new(&path, (size.0, size.1)).into_drawing_area();
 
     let mut cb = ChartBuilder::on(&root_area);
 
@@ -132,7 +132,7 @@ pub(crate) fn pdf_small(
     let path = context.report_path(id, "pdf_small.svg");
 
     let size = size.unwrap_or(SIZE);
-    let root_area = SVGBackend::new(&path, (size.0 as u32, size.1 as u32)).into_drawing_area();
+    let root_area = SVGBackend::new(&path, (size.0, size.1)).into_drawing_area();
 
     let mut chart = ChartBuilder::on(&root_area)
         .margin((5).percent())
@@ -208,7 +208,7 @@ pub(crate) fn pdf(
     let xs_ = Sample::new(&xs);
 
     let size = size.unwrap_or(SIZE);
-    let root_area = SVGBackend::new(&path, (size.0 as u32, size.1 as u32)).into_drawing_area();
+    let root_area = SVGBackend::new(&path, (size.0, size.1)).into_drawing_area();
 
     let range = data::fitting_range(ys.iter());
